+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Forms/ButtonSpec/ButtonSpecFormWindowMax.cs
@@ -10,7 +10,7 @@
         /// Initialize a new instance of the ButtonSpecFormWindowMax class.
         /// </summary>
         /// <param name="form">Reference to owning krypton form instance.</param>
-        public ButtonSpecFormWindowMax(KryptonFormExtended form)
+        public ButtonSpecFormWindowMax(VirtualKryptonFormExtended form)
             : base(form, PaletteButtonSpecStyle.FormMax)
         {
         }
